requirements.txt
torch==2.2.2
torchvision==0.17.2
flask==2.3.3
Pillow==10.0.0
numpy==1.24.3
jaxtyping==0.2.36 